StudentCollaborationDetailService moved to folder Student 	modified:   app/Http/Controllers/api/Student/StudentCollaborationDetailController.php 	renamed:    app/Service/StudentCollaborationDetailService.php -> app/Service/Student/StudentCollaborationDetailService.php 	modified:   tests/Feature/Service/StudentCollaborationDetailServiceTest.php 	modified:   tests/Feature/Student/StudentCollaborationDetailControllerTest.php
